feat: move excepsions and add docstrings
