+++ Trace.Tests/HdrImageTests.cs
@@ -1,6 +1,5 @@
 using System;
 using System.IO;
-using Exceptions;
 using Xunit;
 using Trace;
