+++ Trace.Tests/UtilsTests.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Text;
-using Exceptions;
 using Xunit;
 using Trace;
