+++ Trace/Parameters.cs
@@ -1,5 +1,3 @@
-using Exceptions;
-
 namespace Trace;
 
 using System;
